+++ src/V_fragment/index.test.ts
@@ -3,13 +3,14 @@ import { ctx__new } from '@ctx-core/object'
 import van from 'mini-van-plate/van-plate'
 import { test } from 'uvu'
 import { equal } from 'uvu/assert'
-import { van__set } from '../van'
+import type { PlateElement } from 'van-type-delegate'
 import { V_fragment } from '../index'
+import { van__set } from '../van'
 test('V_fragment', ()=>{
 	const ctx = ctx__new()
 	van__set(ctx, van)
 	const H = van.tags
-	equal(V_fragment(ctx, H.div(H.em('hello!'))).render(),
+	equal((V_fragment(ctx, H.div(H.em('hello!'))) as PlateElement).render(),
 		`<div><em>hello!</em></div>`)
 })
 test.run()