+++ src/V_raw/index.test.ts
@@ -3,11 +3,12 @@ import { test } from 'uvu'
 // @ts-ignore
 import van from 'mini-van-plate/van-plate'
 import { equal } from 'uvu/assert'
+import type { PlateElement } from 'van-type-delegate'
 import { van__set } from '../van'
 import { V_raw } from '../index'
 test('V_raw', ()=>{
 	const ctx = ctx__new()
 	van__set(ctx, van)
-	equal(V_raw(ctx, '<div><em>Hello!</em></div>').render(), '<div><em>Hello!</em></div>')
+	equal((V_raw(ctx, '<div><em>Hello!</em></div>') as PlateElement).render(), '<div><em>Hello!</em></div>')
 })
 test.run()