+++ src/be_state_triple/index.d.ts
@@ -1,20 +1,24 @@
 import type { Be, be__params_T, Ctx } from '@ctx-core/object'
-import type { State } from 'van-type-delegate'
+import type { State, StateVal } from 'van-type-delegate'
 export declare function be_state_triple__new<
-	val_T,
+	state_T extends State<unknown> = State<any>,
 	ctx_T extends Ctx = Ctx
->(...arg_a:be_state_triple__new__arg_a_T<State<val_T>, ctx_T>):be_state_triple_T<State<val_T>, ctx_T>
+>(state__new:(ctx:ctx_T)=>state_T, be__params?:be__params_T):be_state_triple_T<state_T, ctx_T>
+export declare function be_state_triple__new<
+	state_T extends State<unknown> = State<any>,
+	ctx_T extends Ctx = Ctx
+>(id:string|null|undefined, state__new:(ctx:ctx_T)=>state_T, be__params?:be__params_T):be_state_triple_T<state_T, ctx_T>
 export type be_state_triple_T<
-	val_T,
+	state_T extends State<unknown> = State<any>,
 	ctx_T extends Ctx = Ctx
 > = [
-	Be<State<val_T>>,
-	(ctx:ctx_T)=>val_T,
-	(ctx:ctx_T, val:val_T)=>void
+	Be<state_T>,
+	(ctx:ctx_T)=>StateVal<state_T>,
+	(ctx:ctx_T, val:StateVal<state_T>)=>void
 ]
 export type be_state_triple__new__arg_a_T<
-	val_T,
+	state_T extends State<unknown> = State<any>,
 	ctx_T extends Ctx = Ctx
 > =
-	|[state__new:(ctx:ctx_T)=>State<val_T>, be__params?:be__params_T]
-	|[id:string|null|undefined, state__new:(ctx:ctx_T)=>State<val_T>, be__params?:be__params_T]
+	|[state__new:(ctx:ctx_T)=>state_T, be__params?:be__params_T]
+	|[id:string|null|undefined, state__new:(ctx:ctx_T)=>state_T, be__params?:be__params_T]